- fix security session - implement /deposit endpoint - implement /buy endpoint

+++ src/main/java/com/vending/machine/domain/model/User.java
@@ -19,7 +19,7 @@ public class User extends BaseEntity {
     private String password;
     @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<UserRole> roles;
-    private BigDecimal deposit;
+    private Integer deposit;
 
     public void addRole(UserRole userRole) {
         if (roles == null) {
